Setup list in Android
