+++ index.ios.js
@@ -10,40 +10,25 @@ import React, {
   NavigatorIOS,
   StyleSheet,
   Text,
-	ListView,
-  ScrollView,
-  View
+  ListView,
 } from 'react-native';
 
-import DemoSimpleList from './components/DemoSimpleList';
+import RecordListViewIOS from './components/RecordListViewIOS';
 import styles from './styles/Initial';
 import MockData from './data/records';
 
 class Main extends Component {
-	constructor(props) {
-		super(props)
+  constructor(props) {
+    super(props)
     const ds = new ListView.DataSource({rowHasChanged: (r1, r2) => r1 !== r2});
-		this.state = {
-			dataSource: ds.cloneWithRows(MockData.records)
-		}
-	}
-  render() {
-    return (
-      <ScrollView style={styles.list}>
-				<DemoSimpleList dataSource={this.state.dataSource} />
-      </ScrollView>
-    );
+    this.state = {
+      dataSource: ds.cloneWithRows(MockData.records)
+    }
   }
-}
 
-class Add extends Component {
   render() {
     return (
-      <ScrollView style={styles.list}>
-        <Text style={styles.welcome}>
-          Add something here
-        </Text>
-      </ScrollView>
+      <RecordListViewIOS dataSource={this.state.dataSource} />
     );
   }
 }
@@ -52,13 +37,13 @@ class NotitieBlok extends Component {
   render() {
     return (
       <NavigatorIOS
-				style={styles.container}
+        style={styles.container}
         initialRoute={{
           component: Main,
           title: 'Boekingen',
-					leftButtonIcon: require('image!NavBarButtonIcon'),
-					rightButtonIcon: require('image!NavBarButtonPlus'),
-					onLeftButtonPress: () => {console.log('pressed')},
+          leftButtonIcon: require('image!NavBarButtonIcon'),
+          rightButtonIcon: require('image!NavBarButtonPlus'),
+          onLeftButtonPress: () => {console.log('pressed')},
         }}
         itemWrapperStyle={styles.itemWrapper}
       />
